backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
bcrypt==4.1.3
firebase-admin==6.9.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv
Flask-SQLAlchemy==3.1.1
